Code for invitations refactored
Created new helper class for co-author invitation process
